Replace SwinjectStoryboardSpec with SwinjectStoryboardTests.
